Remove services and rename auth feature to user
