+++ src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 import itemReducer from '../features/items/itemsSlice'
-import userReducer from '../features/authentication/authSlice'
+import userReducer from '../features/user/userSlice'
 import dialogReducer from '../features/dialog/dialogSlice'
 import { apiSlice } from '../features/api/apiSlice'
